Uses PreferenceFragment in AdminPreferencesActivity
