+++ collect_app/src/main/java/org/odk/collect/android/preferences/AdminPreferencesActivity.java
@@ -14,33 +14,25 @@
 
 package org.odk.collect.android.preferences;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.ListPreference;
-import android.preference.Preference;
 import android.preference.PreferenceManager;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.view.WindowManager;
-import android.widget.EditText;
 import android.widget.Toast;
 
 import org.javarosa.core.model.FormDef;
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
-import static org.odk.collect.android.preferences.AdminKeys.*;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
+
+import static org.odk.collect.android.preferences.AdminKeys.KEY_FORM_PROCESSING_LOGIC;
 
 /**
  * Handles admin preferences, which are password-protectable and govern which app features and
@@ -50,149 +42,8 @@ import java.io.ObjectOutputStream;
  *         option)
  */
 public class AdminPreferencesActivity extends AppPreferenceActivity {
-
-    public static String ADMIN_PREFERENCES = "admin_prefs";
-
     private static final int SAVE_PREFS_MENU = Menu.FIRST;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setTitle(getString(R.string.admin_preferences));
-
-        PreferenceManager prefMgr = getPreferenceManager();
-        prefMgr.setSharedPreferencesName(ADMIN_PREFERENCES);
-        prefMgr.setSharedPreferencesMode(MODE_WORLD_READABLE);
-
-        addPreferencesFromResource(R.xml.admin_preferences);
-
-        ListPreference mFormProcessingLogicPreference = listPref(KEY_FORM_PROCESSING_LOGIC);
-        mFormProcessingLogicPreference.setSummary(mFormProcessingLogicPreference.getEntry());
-        mFormProcessingLogicPreference.setOnPreferenceChangeListener(
-                new Preference.OnPreferenceChangeListener() {
-
-                    @Override
-                    public boolean onPreferenceChange(Preference preference, Object newValue) {
-                        int index = ((ListPreference) preference).findIndexOfValue(
-                                newValue.toString());
-                        String entry = (String) ((ListPreference) preference).getEntries()[index];
-                        preference.setSummary(entry);
-                        return true;
-                    }
-                });
-
-        Preference mChangeAdminPwPreference = pref(KEY_CHANGE_ADMIN_PASSWORD);
-        mChangeAdminPwPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                AlertDialog.Builder b = new AlertDialog.Builder(AdminPreferencesActivity.this);
-
-                LayoutInflater factory = LayoutInflater.from(AdminPreferencesActivity.this);
-                final View dialogView = factory.inflate(R.layout.password_dialog_layout, null);
-
-                final EditText passwordEditText = (EditText) dialogView.findViewById(R.id.pwd_field);
-                final EditText verifyEditText = (EditText) dialogView.findViewById(R.id.verify_field);
-
-                b.setTitle(R.string.change_admin_password);
-                b.setView(dialogView);
-                b.setPositiveButton(getString(R.string.ok), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        String pw = passwordEditText.getText().toString();
-                        String ver = verifyEditText.getText().toString();
-
-                        if (!pw.equalsIgnoreCase("") && !ver.equalsIgnoreCase("") && pw.equals(ver)) {
-                            // passwords are the same
-                            SharedPreferences.Editor editor = getSharedPreferences(ADMIN_PREFERENCES, MODE_PRIVATE).edit();
-                            editor.putString(KEY_ADMIN_PW, pw);
-                            Toast.makeText(AdminPreferencesActivity.this,
-                                    R.string.admin_password_changed, Toast.LENGTH_SHORT).show();
-                            editor.commit();
-                            dialog.dismiss();
-                            Collect.getInstance().getActivityLogger()
-                                    .logAction(this, "AdminPasswordDialog", "CHANGED");
-                        } else if (pw.equalsIgnoreCase("") && ver.equalsIgnoreCase("")) {
-                            SharedPreferences.Editor editor = getSharedPreferences(ADMIN_PREFERENCES, MODE_PRIVATE).edit();
-                            editor.putString(KEY_ADMIN_PW, "");
-                            editor.commit();
-                            Toast.makeText(AdminPreferencesActivity.this,
-                                    R.string.admin_password_disabled, Toast.LENGTH_SHORT).show();
-                            dialog.dismiss();
-                            Collect.getInstance().getActivityLogger()
-                                    .logAction(this, "AdminPasswordDialog", "DISABLED");
-                        } else {
-                            Toast.makeText(AdminPreferencesActivity.this,
-                                    R.string.admin_password_mismatch, Toast.LENGTH_SHORT).show();
-                            Collect.getInstance().getActivityLogger()
-                                    .logAction(this, "AdminPasswordDialog", "MISMATCH");
-                        }
-                    }
-                });
-                b.setNegativeButton(getString(R.string.cancel), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        dialog.dismiss();
-                        Collect.getInstance().getActivityLogger().logAction(this, "AdminPasswordDialog", "CANCELLED");
-                    }
-                });
-
-                b.setCancelable(false);
-                AlertDialog dialog = b.create();
-                dialog.getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_VISIBLE);
-                dialog.show();
-                return true;
-            }
-        });
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        Collect.getInstance().getActivityLogger()
-                .logAction(this, "onCreateOptionsMenu", "show");
-        super.onCreateOptionsMenu(menu);
-
-        menu
-                .add(0, SAVE_PREFS_MENU, 0, R.string.save_preferences)
-                .setIcon(R.drawable.ic_menu_save)
-                .setShowAsAction(MenuItem.SHOW_AS_ACTION_NEVER);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case SAVE_PREFS_MENU:
-                File writeDir = new File(Collect.SETTINGS);
-                if (!writeDir.exists()) {
-                    if (!writeDir.mkdirs()) {
-                        Toast.makeText(
-                                this,
-                                "Error creating directory "
-                                        + writeDir.getAbsolutePath(),
-                                Toast.LENGTH_SHORT).show();
-                        return false;
-                    }
-                }
-
-                File dst = new File(writeDir.getAbsolutePath()
-                        + "/collect.settings");
-                boolean success = AdminPreferencesActivity.saveSharedPreferencesToFile(dst, this);
-                if (success) {
-                    Toast.makeText(
-                            this,
-                            "Settings successfully written to "
-                                    + dst.getAbsolutePath(), Toast.LENGTH_LONG)
-                            .show();
-                } else {
-                    Toast.makeText(this,
-                            "Error writing settings to " + dst.getAbsolutePath(),
-                            Toast.LENGTH_LONG).show();
-                }
-                return true;
-
-        }
-        return super.onOptionsItemSelected(item);
-    }
+    public static String ADMIN_PREFERENCES = "admin_prefs";
 
     public static boolean saveSharedPreferencesToFile(File dst, Context context) {
         // this should be in a thread if it gets big, but for now it's tiny
@@ -209,8 +60,6 @@ public class AdminPreferencesActivity extends AppPreferenceActivity {
             output.writeObject(adminPreferences.getAll());
 
             res = true;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -268,5 +117,61 @@ public class AdminPreferencesActivity extends AppPreferenceActivity {
         }
 
         return mode;
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getFragmentManager().beginTransaction().replace(android.R.id.content, new AdminPreferencesFragment()).commit();
+        setTitle(getString(R.string.admin_preferences));
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        Collect.getInstance().getActivityLogger()
+                .logAction(this, "onCreateOptionsMenu", "show");
+        super.onCreateOptionsMenu(menu);
+
+        menu
+                .add(0, SAVE_PREFS_MENU, 0, R.string.save_preferences)
+                .setIcon(R.drawable.ic_menu_save)
+                .setShowAsAction(MenuItem.SHOW_AS_ACTION_NEVER);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case SAVE_PREFS_MENU:
+                File writeDir = new File(Collect.SETTINGS);
+                if (!writeDir.exists()) {
+                    if (!writeDir.mkdirs()) {
+                        Toast.makeText(
+                                this,
+                                "Error creating directory "
+                                        + writeDir.getAbsolutePath(),
+                                Toast.LENGTH_SHORT).show();
+                        return false;
+                    }
+                }
+
+                File dst = new File(writeDir.getAbsolutePath()
+                        + "/collect.settings");
+                boolean success = AdminPreferencesActivity.saveSharedPreferencesToFile(dst, this);
+                if (success) {
+                    Toast.makeText(
+                            this,
+                            "Settings successfully written to "
+                                    + dst.getAbsolutePath(), Toast.LENGTH_LONG)
+                            .show();
+                } else {
+                    Toast.makeText(this,
+                            "Error writing settings to " + dst.getAbsolutePath(),
+                            Toast.LENGTH_LONG).show();
+                }
+                return true;
+
+        }
+        return super.onOptionsItemSelected(item);
     }
 }
